todo: - seat connect with date and time, so it can be book correctly

+++ src/main/controller/HistoryController.java
@@ -17,6 +17,10 @@ import main.model.BookCheckingModel;
 import java.io.IOException;
 import java.net.URL;
 import java.sql.*;
+import java.time.LocalDate;
+import java.time.LocalTime;
+import java.time.temporal.ChronoUnit;
+import java.util.Locale;
 import java.util.ResourceBundle;
 
 public class HistoryController implements Initializable
@@ -48,6 +52,7 @@ public class HistoryController implements Initializable
     {
         try
         {
+            autoReject();
             UpdateTable();
         }
         catch (SQLException e)
@@ -75,7 +80,6 @@ public class HistoryController implements Initializable
                 list.add(new HistoryModel(rs.getInt("id"), rs.getString("date"),
                         rs.getString("time"), rs.getString("seat"), rs.getString("approved")));
             }
-
         }
         catch (Exception e)
         {
@@ -116,23 +120,37 @@ public class HistoryController implements Initializable
     }
 
     // delete row from database
-    public void deleteBooking()
+    public void deleteBooking() throws SQLException
     {
+        boolean check = false;
         // need to check first if the selected booking has finished or it's a future one
         if(!bookCheckingModel.checkDateTime(date,time))
         {
             SQLConnection sqlConnection = new SQLConnection();
             Connection connectionDB = sqlConnection.connect();
-
-            String query = "DELETE FROM Booking WHERE id = ?";
-
             PreparedStatement preparedStatement = null;
+            ResultSet rs = null;
+
             try
             {
-                preparedStatement = connectionDB.prepareStatement(query);
+                String sql = "SELECT * FROM Booking WHERE id = ?";
+                preparedStatement = connectionDB.prepareStatement(sql);
                 preparedStatement.setString(1, id);
-                preparedStatement.execute();
-
+                rs = preparedStatement.executeQuery();
+                if (rs.next())
+                {
+                    String r_approved = rs.getString("approved");
+                    if (r_approved.equals("no"))
+                        check = false;
+                    else
+                    {
+                        check = true;
+                        String query = "DELETE FROM Booking WHERE id = ?";
+                        preparedStatement = connectionDB.prepareStatement(query);
+                        preparedStatement.setString(1, id);
+                        preparedStatement.execute();
+                    }
+                }
                 UpdateTable();
             }
             catch (Exception e)
@@ -140,8 +158,20 @@ public class HistoryController implements Initializable
                 e.printStackTrace();
                 e.getCause();
             }
-            failMessage.setText("");
-            successMessage.setText("Cancel booking successfully !");
+            finally
+            {
+                preparedStatement.close();
+                rs.close();
+                connectionDB.close();
+            }
+            if (check)
+            {
+                failMessage.setText("");
+                successMessage.setText("Cancel booking successfully !");
+            }
+            else
+                successMessage.setText("");
+                failMessage.setText("Cannot cancel booking that is not be approved.");
         }
         else
         {
@@ -150,8 +180,63 @@ public class HistoryController implements Initializable
         }
     }
 
+    //auto reject booking if admin do not respond 1 hour before the time
+    public void autoReject() throws SQLException
+    {
+        //get connection from database to table
+        SQLConnection sqlConnection = new SQLConnection();
+        Connection connectionDB = sqlConnection.connect();
+        PreparedStatement preparedStatement = null;
+        ResultSet rs = null;;
+
+        try
+        {
+            preparedStatement = connectionDB.prepareStatement("SELECT * FROM Booking;");
+            rs = preparedStatement.executeQuery();
+
+            while (rs.next())
+            {
+                String r_approved = rs.getString("approved").toLowerCase();
+                if (r_approved.equals("pending"))
+                {
+                    int r_id = rs.getInt("id");
+                    LocalDate r_date = LocalDate.parse(rs.getString("date"));
+                    LocalTime r_time = LocalTime.parse(rs.getString("time"));
+
+                    LocalDate localDate = LocalDate.now();
+                    LocalTime localTime = LocalTime.now();
+
+                    int dateDiff = r_date.compareTo(localDate);
+                    if (dateDiff == 0)
+                    {
+                        int hourBeforeCancel = 1;
+                        long hoursBetween = ChronoUnit.HOURS.between(localTime, r_time);
+
+                        if (hoursBetween < hourBeforeCancel)
+                        {
+                            String sql = "UPDATE Booking SET approved = '" + "no(auto cancel)" + "'WHERE id = '" + r_id + "';";
+                            preparedStatement = connectionDB.prepareStatement(sql);
+                            preparedStatement.execute();
+                        }
+                    }
+                }
+            }
+        }
+        catch (Exception e)
+        {
+            e.printStackTrace();
+            e.getCause();
+        }
+        finally
+        {
+            connectionDB.close();
+            preparedStatement.close();
+            rs.close();
+        }
+    }
+
     //cancel booking button action event
-    public void Cancel(ActionEvent event) throws IOException
+    public void Cancel(ActionEvent event) throws IOException, SQLException
     {
         deleteBooking();
     }
